# api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Reportnet3 Dataflow API",
    "version": "1.0"
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header. A valid JWT token is generated by the API and retured to the caller after a successful call to the 'login' route. The following syntax must be used in the 'Authorization' header : Bearer: followed by the space and the JWT token."
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
    "/api/auth/signin": {
      "post": {
        "summary": "Sign in",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid username or password"
          }
        }
      }
    },
    "/api/dataflow/reportnet3/csv": {
      "get": {
        "summary": "Get dataflows in csv format",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "403": {
            "description": "Reportnet3 dataflow is not enabled"
          },
          "400": {
            "description": "Could not create csv for dataflow, no data found"
          }
        }
      }
    },
    "/api/dataflow/reportnet3/b": {
      "get": {
        "summary": "Get dataflow b",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "403": {
            "description": "Reportnet3 dataflow is not enabled"
          },
          "400": {
            "description": "No data found"
          }
        }
      }
    },
    "/api/dataflow/reportnet3/d/processes": {
      "get": {
        "summary": "Get dataflow d processes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "403": {
            "description": "Reportnet3 dataflow is not enabled"
          },
          "400": {
            "description": "No data found"
          }
        }
      }
    },
    "/api/dataflow/reportnet3/d/samplingpoints": {
      "get": {
        "summary": "Get dataflow d sampling points",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "403": {
            "description": "Reportnet3 dataflow is not enabled"
          },
          "400": {
            "description": "No data found"
          }
        }
      }
    }
  }
}